비동기 api 데이터 타입 재사용 로직으로 분리
