+++ src/types/profile.ts
@@ -1,0 +1,9 @@
+export interface Profile {
+  name: string;
+  email: string;
+  tel: string;
+  // field: string;
+  // job: string;
+  introduce: string;
+  myKeywords: string[] | [];
+}
